Add `watch=true` option to the REST API
The option allows all REST API resources that support the GET method to be
accessed via WebSockets. The requested resource is written into the open
socket every time it changes. The method of detecting changes is still
based on polling so it is not horribly efficient nor is it fast.