Add unique message ID
The journal entry cursor position is used as the ID when messages are
extracted from the systemd journal. For files, the only unique identifier
we have is the line number.

Also removed the appending of messages as it caused confusing messages to
be printed when a log delimiter line was encountered.